requirements.txt
python-telegram-bot==20.6
yt-dlp==2026.3.17
requests==2.32.0
instaloader==4.12